#6519 – Combine accordion and tabs
