+++ packages/ketcher-macromolecules/src/components/monomerLibrary/RnaBuilder/RnaAccordion/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  ***************************************************************************/
 
-export { RnaAccordion } from './RnaAccordion';
+export { RnaUnifiedView } from './RnaUnifiedView';
